.vimrc: Update C++ plugins with LSP alternatives
vim-clang-format and others were necessary before LSP was as mature as
it now is, however LSP is quite good now so prefer to use it where
possible.

There is still room for improvement on the LSP configuration in terms of
keybinds and probably other settings, e.g. snippets, but this should be
a decent starting point.